feat: Fix rebase, rename file CY-7427
Renamed file to better fit rest of the project.
