+++ src/tools/index.ts
@@ -17,5 +17,5 @@ export * from './listRepositoryToolPatternsTool.js';
 export * from './listRepositoryToolsTool.js';
 export * from './listToolsTool.js';
 export * from './installCLITool.js';
-export * from './CLIanalysisTool.js';
+export * from './cliAnalysisTool.js';
 export * from './listOrganizationsTool.js';
